--- OrangeBricks.Web/Controllers/Property/Commands/BookViewingCommand.cs
@@ -1,16 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Web;
-
-namespace OrangeBricks.Web.Controllers.Property.Commands
-{
-    public class BookViewingCommand
-    {
-        public int PropertyId { get; set; }
-
-        public DateTime ViewingDateTime { get; set; }
-
-        public string BuyerUserId { get; set; }
-    }
-}
