+++ OrangeBricks.Web/Controllers/Property/PropertyController.cs
@@ -94,42 +94,6 @@ namespace OrangeBricks.Web.Controllers.Property
             return RedirectToAction("Index");
         }
 
-        [OrangeBricksAuthorize(Roles = "Buyer")]
-        public ActionResult BookViewing(int id)
-        {
-            var builder = new BookViewingViewModelBuilder(_context);
-            var viewModel = builder.Build(id);
-            return View(viewModel);
-        }
-
-        [HttpPost]
-        [OrangeBricksAuthorize(Roles = "Buyer")]
-        public ActionResult BookViewing(BookViewingCommand command)
-        {
-            var handler = new BookViewingCommandHandler(_context);
-
-            command.BuyerUserId = User.Identity.GetUserId();
-
-            handler.Handle(command);
-            return RedirectToAction("Index");
-        }
-
-        [OrangeBricksAuthorize(Roles = "Seller")]
-        public ActionResult ViewingsBooked(int id)
-        {
-            var builder = new BookedViewingsViewModelBuilder(_context);
-            var viewModel = builder.Build(id);
-
-            return View(viewModel);
-        }
-
-        [OrangeBricksAuthorize(Roles = "Buyer")]
-        public ActionResult MyViewings()
-        {
-            var builder = new MyViewingsViewModelBuilder(_context);
-            var viewModel = builder.Build(User.Identity.GetUserId());
-
-            return View(viewModel);
-        }
+        
     }
 }
