+++ OrangeBricks.Web/Controllers/Viewing/Builders/BookViewingViewModelBuilder.cs
@@ -1,11 +1,8 @@
-﻿using OrangeBricks.Web.Controllers.Property.ViewModels;
+﻿using OrangeBricks.Web.Controllers.Viewing.ViewModels;
 using OrangeBricks.Web.Models;
 using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Web;
 
-namespace OrangeBricks.Web.Controllers.Property.Builders
+namespace OrangeBricks.Web.Controllers.Viewing.Builders
 {
     public class BookViewingViewModelBuilder
     {
