+++ OrangeBricks.Web/Controllers/Viewing/Builders/MyViewingsViewModelBuilder.cs
@@ -1,9 +1,9 @@
-﻿using OrangeBricks.Web.Controllers.Property.ViewModels;
+﻿using OrangeBricks.Web.Controllers.Viewing.ViewModels;
 using OrangeBricks.Web.Models;
 using System.Collections.Generic;
 using System.Linq;
 
-namespace OrangeBricks.Web.Controllers.Property.Builders
+namespace OrangeBricks.Web.Controllers.Viewing.Builders
 {
     public class MyViewingsViewModelBuilder
     {
@@ -25,7 +25,7 @@ namespace OrangeBricks.Web.Controllers.Property.Builders
             };
         }
 
-        private IEnumerable<BookViewingViewModel> CreateViewingViewModel(List<Viewing> viewings)
+        private IEnumerable<BookViewingViewModel> CreateViewingViewModel(List<Models.Viewing> viewings)
         {
             List<BookViewingViewModel> bookViewings = new List<BookViewingViewModel>();
