+++ OrangeBricks.Web/Controllers/Viewing/Commands/BookViewingCommand.cs
@@ -1,0 +1,11 @@
+﻿using System;
+
+namespace OrangeBricks.Web.Controllers.Viewing.Commands
+{
+    public class BookViewingCommand
+    {
+        public int PropertyId { get; set; }
+        public DateTime ViewingDateTime { get; set; }
+        public string BuyerUserId { get; set; }
+    }
+}
