+++ OrangeBricks.Web/Models/Viewing.cs
@@ -1,8 +1,5 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
-using System.Linq;
-using System.Web;
 
 namespace OrangeBricks.Web.Models
 {
