+++ OrangeBricks.Web/Views/Property/Index.cshtml
@@ -30,7 +30,7 @@
     <p>@property.NumberOfBedrooms bedroom @property.PropertyType</p>
     <p>@property.Description</p>
 
-    <a href="@Url.Action("BookViewing","Property", new {id = property.Id })" class="btn btn-primary">
+    <a href="@Url.Action("BookViewing","Viewing", new {id = property.Id })" class="btn btn-primary">
         Book Viewing
     </a>
     <a href="@Url.Action("MakeOffer", "Property", new {id = property.Id})" class="btn btn-primary">
