+++ OrangeBricks.Web/Views/Shared/_Layout.cshtml
@@ -33,7 +33,7 @@
                         if (User.IsInRole("Buyer"))
                         {
                             <li>@Html.ActionLink("My Offers","MyOffers","Offers")</li>
-                            <li>@Html.ActionLink("My Viewings", "MyViewings", "Property")</li>
+                            <li>@Html.ActionLink("My Viewings", "MyViewings", "Viewing")</li>
                         }
 
                     }
