+++ OrangeBricks.Web/Views/Viewing/MyViewings.cshtml
@@ -1,4 +1,4 @@
-﻿@model OrangeBricks.Web.Controllers.Property.ViewModels.MyViewingsViewModel
+﻿@model OrangeBricks.Web.Controllers.Viewing.ViewModels.MyViewingsViewModel
 @{
     ViewBag.Title = "Viewings booked";
     Layout = "~/Views/Shared/_Layout.cshtml";
