+++ OrangeBricks.Web/Views/Viewing/ViewingsBooked.cshtml
@@ -1,4 +1,4 @@
-﻿@model OrangeBricks.Web.Controllers.Property.ViewModels.BookedViewingsViewModel
+﻿@model OrangeBricks.Web.Controllers.Viewing.ViewModels.BookedViewingsViewModel
 @{
     ViewBag.Title = "Viewings for you Property";
     Layout = "~/Views/Shared/_Layout.cshtml";
